Final push, new contract deploy and resolvers implemented
