+++ src/components/resolvers/Rinkeby/index.js
@@ -1,1 +1,1 @@
-export default ["0x16fD6e2E1C4afB9C4e7B901141706596317e4ceB","0xa8033faC7cC8E8f650e1405493D4f1317E7dE9BB"]
+export default ["0x16fD6e2E1C4afB9C4e7B901141706596317e4ceB","0x070a6EA824d87780643E888196CD92d748B02A22"]
